src/slash_bot_client/requirements.txt
discord-py-interactions==4.3.4
numpy==1.21.5
opencv-python==4.5.4.60
pandas==1.3.5
pika==1.3.1
psycopg2-binary==2.9.3
scikit-learn==1.0.2
SQLAlchemy==1.4.33
nest-asyncio==1.5.4